Update p3.java
double to float

+++ p3.java
@@ -4,7 +4,7 @@ public class p3 {
     public static void main(String[] args) {
         Scanner s = new Scanner(System.in);
         int yearOfSrevice;
-        double salary, bonusPercent = 0.05;
+        float salary, bonusPercent = 0.05;
         System.out.println("Enter your Salary : ");
         salary = s.nextInt();
         System.out.println("Enter your year of service i this company : ");
